apps: zynqmp: move rproc implementation to separate file.
Previously, linux to r5 rproc implementatio was in platform info.
Mimic zynq7 implementation to move rproc ops to separate file.
