+++ apps/machine/zynqmp/CMakeLists.txt
@@ -1,3 +1,4 @@
 collect (APP_COMMON_SOURCES platform_info.c)
+collect (APP_COMMON_SOURCES zynqmp_linux_r5_proc.c)
 collect (APP_INC_DIRS "${CMAKE_CURRENT_SOURCE_DIR}")
